// src/components/atoms/enum.tsx
export enum Variant {
  Solid = 'solid',
  Bordered = 'bordered',
  Light = 'light',
  Flat = 'flat',
  Faded = 'faded',
  Shadow = 'shadow',
  Ghost = 'ghost',
}

export enum Color {
  Default = 'default',
  Primary = 'primary',
  Secondary = 'secondary',
  Success = 'success',
  Error = 'error',
  Warning = 'warning',
  Danger = 'danger',
}

export enum Size {
  Small = 'sm',
  Medium = 'md',
  Large = 'lg',
}

export enum Radius {
  None = 'none',
  Small = 'sm',
  Medium = 'md',
  Large = 'lg',
  Full = 'full',
}

export enum SpinnerPlacement {
  Start = 'start',
  End = 'end',
}

export enum AccordionVariant {
  Light = 'light',
  Shadow = 'shadow',
  Bordered = 'bordered',
  Splitted = 'splitted',
}

export enum SelectionMode {
  Single = 'single',
  Multiple = 'multiple',
  None = 'none',
}

export enum SelectionBehavior {
  Toggle = 'toggle',
  Replace = 'replace',
}